+++ backend/apps/slack/common/handlers/calendar_events.py
@@ -62,7 +62,6 @@ def get_reminder_blocks(args, slack_user_id: str) -> list[dict]:
     """Get the blocks for setting a reminder."""
     from apps.nest.controllers.calendar_events import set_reminder
     from apps.nest.schedulers.calendar_events.slack import SlackScheduler
-    from apps.nest.schedulers.tasks.slack import send_message
 
     try:
         reminder_schedule = set_reminder(
@@ -75,7 +74,7 @@ def get_reminder_blocks(args, slack_user_id: str) -> list[dict]:
         )
         scheduler = SlackScheduler(reminder_schedule)
         scheduler.schedule()
-        send_message(
+        SlackScheduler.send_message(
             f"<@{reminder_schedule.reminder.member.slack_user_id}> set a reminder: "
             f"{reminder_schedule.reminder.message}"
             f" at {reminder_schedule.scheduled_time.strftime('%Y-%m-%d %H:%M %Z')}",
